aihero/course/pyproject.toml
[project]
name = "course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
